add input element
