+++ src/core/events/change.event.ts
@@ -1,0 +1,6 @@
+export class UmbChangeEvent extends Event {
+	public constructor() {
+		// mimics the native change event
+		super('change', { bubbles: true, composed: false, cancelable: false });
+	}
+}
